+++ templates/index.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+  </head>
+  <body>
+    <div class="wrapper">
+      <header class="title">
+        <h1>台北一日遊</h1>
+        <div class="nav">
+          <div class="nav-item">
+            <a href="">預定行程</a>
+          </div>
+          <div class="signin">
+            <p>登入/註冊</p>
+          </div>
+        </div>
+      </header>
+      <section class="hero">
+        <h2>輕鬆享受台北一日遊</h2>
+        <div class="slogan">
+          <p>探索每個角落，體驗城市的深度旅遊行程</p>
+          <form action="">
+            <input type="text" name="" id="" placeholder="輸入景點名稱查詢" />
+            <button></button>
+          </form>
+        </div>
+      </section>
+      <main>
+        <div class="mrt-ctn">
+          <div class="mrt-ite"></div>
+        </div>
+        <div class="attractions-ctn">
+          <div class="attraction-item">
+            <div class="item-pic">
+              <img src="" alt="" />
+            </div>
+            <div class="item-name">
+              <p></p>
+            </div>
+            <div class="item-details">
+              <p>忠孝復興</p>
+              <p>公共藝術</p>
+            </div>
+          </div>
+        </div>
+      </main>
+      <footer>
+        <p>COPYRIGHT © 2021 台北一日遊</p>
+      </footer>
+    </div>
+  </body>
+</html>
